+++ SpellPouch/changelog.md
@@ -1,4 +1,4 @@
 # Changelog
 
 ## [1.0.0]
-- first release
+- added ability/spell groups